Extract sidebar and topbar in components
