+++ client/pages/configuration.vue
@@ -1,13 +1,12 @@
 <script lang="ts" setup>
 const { t } = useI18n()
 
+usePageName(t('pages.configuration.page_name'))
 definePageMeta({
   layout: 'sidebar',
 })
 </script>
 
 <template>
-  <h2 class="text-3xl" style="margin: .375rem 0;">
-    {{ t('sidebar.configuration') }}
-  </h2>
+  <p>TODO</p>
 </template>
